pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "meeting-transcriber"
version = "0.1.0"
description = "Offline Korean meeting transcriber and translator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faster-whisper>=1.0.0",
    "transformers>=4.36.0",
    "sentencepiece>=0.1.99",
    "ffmpeg-python>=0.2.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
gpu = ["torch>=2.0.0"]

[project.scripts]
meeting-transcriber = "meeting_transcriber.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/meeting_transcriber"]

